added docker tests for token/player repo fixed methods

+++ internal/models/player.go
@@ -4,6 +4,6 @@ type Player struct {
 	ID          uint64 `json:"id"`
 	Username    string `json:"username"`
 	PhoneNumber string `json:"phone_number"`
-	ChatID      int    `json:"chat_id"`
+	ChatID      int64  `json:"chat_id"`
 	CreatedAt   int64  `json:"created_at"`
 }
