+++ internal/models/player_find_options.go
@@ -3,5 +3,5 @@ package models
 type PlayerFindOptions struct {
 	ID          uint64
 	PhoneNumber string
-	ChatID      int
+	ChatID      int64
 }
